% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gray_bars.R
\name{gen_gray_bars}
\alias{gen_gray_bars}
\title{internal helper function to create gray portion of packed barchart}
\usage{
gen_gray_bars(summ_dt, number_rows, color_bar_data, label_column,
  min_label_width)
}
\description{
internal helper function to create gray portion of packed barchart
}
\keyword{internal}
