% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plotly_packed_bar.R
\name{gen_plotly_packed_bar}
\alias{gen_plotly_packed_bar}
\title{internal helper function to do actual plotting of packed barchart}
\usage{
gen_plotly_packed_bar(summ_dt, value_column, color_bar_color, hover_label,
  canvas_df, plot_title, xaxis_label, gray_hover_point_dt,
  colored_hover_point_dt, gray_bar_list, colored_bar_list, gray_ann_list,
  colored_ann_list, tick_breaks, tick_text)
}
\description{
internal helper function to do actual plotting of packed barchart
}
\keyword{internal}
