% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_xaxis_labels.R
\name{gen_xaxis_labels}
\alias{gen_xaxis_labels}
\title{internal helper function to create xaxis labels}
\usage{
gen_xaxis_labels(row_sums, total_x)
}
\description{
internal helper function to create xaxis labels
}
\keyword{internal}
