% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_bar_count.R
\name{guess_bar_count}
\alias{guess_bar_count}
\title{internal function to 'guess' correct number of colored bars in packed barchat using an elbow method}
\usage{
guess_bar_count(num_vec, min_bar = 3, max_bar = 25)
}
\description{
internal function to 'guess' correct number of colored bars in packed barchat using an elbow method
}
\keyword{internal}
